Refactor: Custom Commands for degoogled and Tempus flavors (#641)
Refactor custom commands
